requirements.txt
aiogram==3.2.0
asyncpg==0.29.0
aiohttp==3.9.1
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.2
redis==5.0.1
cryptography==41.0.8
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
